Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cassandra-driver>=3.29.2",
    "fastapi[standard]>=0.115.5",
    "jwt>=1.3.1",
    "pyjwt>=2.10.0",
]
